Detection hit a roadblock error that I do not understand. Copied predictor as base learner and redoing the refactor
