+++ multimodal/src/autogluon/multimodal/learners/__init__.py
@@ -1,2 +1,3 @@
 from .default_learner import DefaultLearner
 from .object_detection_learner import ObjectDetectionLearner
+from .base_learner import BaseLearner
