+++ multimodal/src/autogluon/multimodal/utils/inference.py
@@ -475,7 +475,6 @@ def predict(
             data=data,
             requires_label=requires_label,
         )
-
     strategy = "dp"  # default used in inference.
 
     num_gpus = compute_num_gpus(config_num_gpus=predictor._config.env.num_gpus, strategy=strategy)
